Improved lua_tests.cpp to handle running lua-based test suites better   Individual lua tests count towards the unittest++ test totals, and report failures individually Wrote some preliminary opengl tests
git-svn-id: http://fracture.googlecode.com/svn/trunk@9 1953d6fe-3836-0410-a91e-07494a8b0067

+++ ext/luabind/object.hpp
@@ -227,7 +227,7 @@ LUABIND_BINARY_OP_DEF(<, lua_lessthan)
         , static_cast<ValueWrapper const&>(v));
 		char const* p = lua_tostring(interpreter, -1);
 		if (p) {
-  		int len = lua_strlen(interpreter, -1);
+  		size_t len = lua_strlen(interpreter, -1);
 	  	std::copy(p, p + len, std::ostream_iterator<char>(os));
 		  return os;
 		} else {
